fix, improve & translate EventFragment

+++ js/src/forum/Components/EventFragment.js
@@ -1,11 +1,10 @@
 import Component from 'flarum/common/Component';
+import Alert from 'flarum/common/components/Alert';
 import Button from 'flarum/common/components/Button';
+import Link from 'flarum/common/components/Link';
 import username from 'flarum/helpers/username';
-import User from 'flarum/models/User'
 import userOnline from 'flarum/helpers/userOnline';
-import avatar from 'flarum/helpers/avatar';
 import EditEventModal from "./EditEventModal";
-import Alert from 'flarum/common/components/Alert'
 import fullTime from 'flarum/helpers/fullTime';
 
 export default class EventFragment extends Component {
@@ -25,13 +24,13 @@ export default class EventFragment extends Component {
   view() {
       return <div>
         <p id="eventdescription"/>
-        <p>Hosted by: <a href={app.route.user(this.attrs.event.user())} config={m.route}>
+        <p>{app.translator.trans('flarum-calendar.forum.event.hosted_by')} <Link href={app.route.user(this.attrs.event.user())}>
           {userOnline(this.attrs.event.user())}
           {username(this.attrs.event.user())}
-        </a></p>
+        </Link></p>
         <p>
-          Starts: { fullTime(this.attrs.event.event_start()) } <br/>
-          Ends: { fullTime(this.attrs.event.event_end())}
+          {app.translator.trans('flarum-calendar.forum.event.starts')} { fullTime(this.attrs.event.event_start()) } <br/>
+          {app.translator.trans('flarum-calendar.forum.event.ends')} { fullTime(this.attrs.event.event_end())}
         </p>
         {(app.session.user && (app.session.user.canModerateEvents || this.attrs.event.user.id === app.session.user.id)) ?
           (<div>
@@ -39,7 +38,7 @@ export default class EventFragment extends Component {
                 icon: 'fas fa-edit',
                 onclick: this.editLaunch.bind(this),
                 className: 'Button Button--icon Button--link',
-              })},
+              })}
               {Button.component({
                 icon: 'fas fa-trash-alt',
                 onclick: this.deleteEvent.bind(this),
@@ -64,11 +63,17 @@ export default class EventFragment extends Component {
   }
 
   deleteEvent() {
+    if (!confirm(app.translator.trans('flarum-calendar.forum.event.confirm_delete'))) {
+      return;
+    }
     // console.log({"message": "[webbinaro/flarum-calendar] delete event ", "event": this.attrs.event})
     const events = this.attrs.events;
     this.attrs.event.delete().then(()=>{
-      app.alerts.show("Event Deleted");
-      m.route(app.route('advevents'));
+      app.alerts.show(Alert, {type: 'success'}, app.translator.trans('flarum-calendar.forum.event.deleted'));
+      m.route.set(app.route('advevents'));
+      if (this.attrs.modal) {
+        this.attrs.modal.hide();
+      }
       //app.history.back();
     });
   }
